+++ lib/map_controller_plus.dart
@@ -2,6 +2,3 @@ library map_controller_plus;
 
 export 'src/controller.dart';
 export 'src/models.dart';
-export 'src/tile_layer/layers.dart';
-export 'src/tile_layer/tile_layers_bar.dart';
-export 'src/types.dart';